Added build instructions to readme, tweaked composer.json, and removed sculpin.lock from gitignore again.
If you are going to have a sculpin.json at all, you should include sculpin.lock
in your repository. Since sculpin.json is more or less empty, no sculpin.lock is
currently created. Whether the server CAN generate sculpin.lock on its own or
not is besides the point. Much like composer.lock, it is best if your dev
environment and your production environments are exactly the same. That is what
sculpin.lock is for.

I originally removed sculpin.json because there was no need for it. If you'd
like to have it anyway, that is fine. I cleaned it up and removed some bits that
don't make sense. If you want to pick a license for this website, you should
pick one other than MIT, I think? Pick a license more suited for content.

I think that in the past I've advocated for not having more or less empty
sculpin.json but when I considered adding the "default" build instructions to
the README, it made me realize that build instructions that work for EVERY
sculpin site would require that EVERY sculpin site have at least a sculpin.json.
So I added it back.

This has the side effect of making your build instructions future proof. You
don't have extensions or anything installed right now, but if you do in the
future you won't have to update your build instructions.